pyproject.toml
[project]
name = "wayfarer"
version = "0.1.0"
description = "Wayfarer: a distributed hardware bridge (CAN, MAVLink ⇄ MQTT), device-first, discovery-driven."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [{name="Your Team"}]
dependencies = [
  "pymavlink>=2.4",
  "paho-mqtt>=2.0",
  "PyYAML>=6.0"
]

[project.scripts]
wayfarer = "wayfarer.cli.main:main"
